adiciona rate limiter

+++ pkg/ratelimiter/cache/cache.go
@@ -3,7 +3,7 @@ package cache
 import "context"
 
 type Cache interface {
-	Get(context.Context, string) (int64, error)
+	Get(context.Context, string) (int, error)
 	Incr(context.Context, string) error
-	Expire(context.Context, string, int64) error
+	Expire(context.Context, string, int) error
 }
